self review code finished
