--- pkg/sink/kafka/error_event.go
@@ -1,39 +1,0 @@
-// Copyright 2023 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package kafka
-
-import (
-	"github.com/IBM/sarama"
-)
-
-func isTransientKError(err *sarama.ProducerError) bool {
-	if err == nil {
-		return false
-	}
-	kerr, ok := err.Err.(sarama.KError)
-	if !ok {
-		return false
-	}
-	switch kerr {
-	case sarama.ErrNotEnoughReplicasAfterAppend,
-		sarama.ErrNotEnoughReplicas,
-		sarama.ErrLeaderNotAvailable,
-		sarama.ErrNotLeaderForPartition,
-		sarama.ErrRequestTimedOut,
-		sarama.ErrNetworkException:
-		return true
-	default:
-	}
-	return false
-}
