--- pkg/sink/recoverable/error_event.go
@@ -1,32 +1,0 @@
-// Copyright 2026 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package recoverable
-
-import (
-	"time"
-
-	"github.com/pingcap/ticdc/pkg/common"
-)
-
-// RecoverEvent is a structured recover event notification emitted by a sink.
-// DispatcherManager forwards it to maintainer to trigger localized recovery.
-type RecoverEvent struct {
-	Time          time.Time
-	DispatcherIDs []common.DispatcherID
-}
-
-// Recoverable is an optional interface implemented by sinks that can report recoverable request.
-type Recoverable interface {
-	SetRecoverEventCh(ch chan<- *RecoverEvent)
-}
